Address zemao's comments.
